Rename fader files to Fader
Renamed the files with a capital F so they properly mirror the name of the class.
